Cups can be dragged and released
The cups can be dragged and released to multiple areas (coffee machine, customer area, coffee preparing area, garbage can). Added a preparing area so cups can be placed there once they are out of the starting position (counter shelve). The cups have the functionality and setup to do the different combinations of coffees